Move all image route tests to same file.
