+++ tests/view_tests/test_views.py
@@ -20,8 +20,6 @@ def client():
         yield client
         subprocess.call(["rm","elephant_vending_machine/static/logs/test_file.csv"])
         subprocess.call(["rm","elephant_vending_machine/static/logs/test_file2.csv"])
-        subprocess.call(["rm", "elephant_vending_machine/static/img/test_file.png"])
-        subprocess.call(["rm", "elephant_vending_machine/static/img/test_file2.jpg"])
 
 def test_run_trial_route_success(client, monkeypatch):
     monkeypatch.setattr('elephant_vending_machine.views.create_experiment_logger', lambda file_name: MockLogger())
@@ -40,14 +38,5 @@ def test_get_log_endpoint(client):
     response = client.get('/log')
     response_json_files = json.loads(response.data)['files']
     min_elements_expected = ["http://localhost/static/logs/test_file.csv","http://localhost/static/logs/test_file2.csv","http://localhost/static/logs/unittest.csv"]
-    assert all(elem in response_json_files for elem in min_elements_expected)
-    assert response.status_code == 200
-
-def test_get_image_endpoint(client):
-    subprocess.call(["touch", "elephant_vending_machine/static/img/test_file.png"])
-    subprocess.call(["touch", "elephant_vending_machine/static/img/test_file2.jpg"])
-    response = client.get('/image')
-    response_json_files = json.loads(response.data)['files']
-    min_elements_expected = ["http://localhost/static/img/test_file.png","http://localhost/static/img/test_file2.jpg"]
     assert all(elem in response_json_files for elem in min_elements_expected)
     assert response.status_code == 200
